L20 Authorization using gates

+++ app/Http/Controllers/QuestionsController.php
@@ -87,6 +87,11 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question) // Route Model Binding
     {
+        //if (\Gate::allows('update-question', $question) === false) {
+        if (\Gate::denies('update-question', $question)) {
+            return abort(403, 'Access denied!');
+        }
+
         return view('questions.edit', compact('question'));
     }
 
@@ -99,6 +104,10 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        if (\Gate::denies('update-question', $question)) {
+            return abort(403, 'Access denied!');
+        }
+
         $question->update($request->only('title', 'body'));
 
         return redirect()->route('questions.index')->with('success', 'Your question has been modified.');
@@ -112,6 +121,10 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        if (\Gate::denies('delete-question', $question)) {
+            return abort(403, 'Access denied!');
+        }
+
         $question->delete();
 
         return redirect()->route('questions.index')->with('success', 'Your question has been removed.');
